Refactor episode card
